MoneyMind_Backend/requirements.txt
# Existing requirements
google-generativeai>=0.3.0
python-dotenv>=1.0.0
serpapi>=0.1.0
fastapi>=0.103.0
google-search-results>=2.4.0
requests>=2.28.0
Flask-Cors>=3.0.10
typing-extensions>=4.5.0
pytest>=7.0.0

# Database dependencies
sqlalchemy>=2.0.12
psycopg2-binary>=2.9.6
alembic>=1.10.4

# Additional dependencies
python-multipart>=0.0.6
uvicorn>=0.22.0
twilio>=9.0.0
pyngrok>=7.0.0
pytesseract>=0.3.10
Pillow>=9.0.0
grpcio>=1.62.0
protobuf==4.25.5 